Added terrain from heightmap, updated controls
Re-did the terrain - it now has a mesh and collision body
  - The bulk of the work is done by terrain_generation.gd (attached to TerrainMesh)
    - It creates and modifies ArrayMesh according to the data from terrain_heightmap_smooth.png
    - The dimensions of the terrain are set to be the size of the png. Note: Only compatible with square images for now.
    - It generates normals based on the positions of the vertices. The normalmap.png is not used.
    - It runs a few passes of very crude smoothing to reduce blocky look of terrain. It looks OK for now but could use improvement.
    - A heightmap array is created so that CollisionShape3D.gd can access it later
  - CollisionShape3D.gd just looks at the heightmap array generated by terrain_generation.gd and creates a HeightMapShape3D collision shape that matches the mesh.

The terrain is currently a child of a NavigationRegion3D (with an attached mostly blank script). For now this does nothing, but it looks like it might be needed later to set up pathing.

Changed controls:
  - WASD now also moves the view around the same way the arrow keys did before
  - The mouse scroll wheel zooms the view in and out. Not sure if we want this to be in the final game but it's REALLY nice to have for testing

I had created this new terrain in a separate scene for testing, but now that it is complete I changed world.tscn to have the new terrain and moved the old scene to world_old.tscn just in case we need anything from it. However, world_old.tscn is a bit broken because I had to change terrain_heightmap_smooth.png to import as Image instead of Texture2D.